Changing logic for how metal changes color from being heated.
The metals emissive color changes based on the temperature of the metal. Original logic: When the metal's temperature reaches a certain value the emissive color of the metal bars material is hard set. New logic: A color gradient of a metal incandescent is hard coded at start time. In the MetalBarController component of each metal bar the color of the metal bar's emissive material is set to a color on the gradient at a given time. The time of the color lies between [0,1]. A function is created so that given the temperature of the metal bar (in the range of incandescent light) it will return a float that lies in the range [0,1].
This results in a more gradual transition between the colors., also  less LOC in the Update function.