Heating/Cooling implement refined.
Trigger mechanic for H/C changed so that no for loop is used to see if forge/cooler has an object to heat/cool. A global timer is also being implemented so that the tracking and triggering per second used for H/C  is limited to one timer instead of one timer for each H/C'd instantiation.