Changing Item PickUp Mechanics
Instead of holding object constantly changing pos + rot to match player; item is placed in holding pos and then made a child of the tool used to carry it (also within player GO hierarchy).  Advantage is that pos and rot of holding obj is set once vs const updated when being carried.