Malleability + Shanpe Changing Script implemented
Malleability script attached as a component to the gameobject that holds the circlecollider at the surface of the hammer head. Change Shape script added to metal bards. When the circle collider on the hammer collides with the metal bar it triggers ChangeShape func within Change Shape component.  ChangeShape func takes in stress hammerhead produces and the direction the hammer head is facing (vector3d). Both parameters determine how much change in the metal bars shape and which 2 axis are elongated and which one is compressed.  Change in Shape mechanic is not finished